requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
dotenv
pydantic